created top panel container to pass props to component
