+++ web-client/src/modules/timeline/pages/ContentPage.tsx
@@ -1,6 +1,7 @@
 import React, { ReactElement } from 'react';
-import TopPanel from 'src/components/RequestsView/TopPanel';
 
-const ContentPage = (): ReactElement => <TopPanel />;
+import TopPanelContainer from '../containers/TopPanelContainer';
+
+const ContentPage = (): ReactElement => <TopPanelContainer />;
 
 export default ContentPage;
